backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2",
    "django-ninja>=1.4.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.3",
    "mypy>=1.15.0",
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]  # E(PEP8), F(잠재적 오류), I(Import 정리-isort)
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
mypy_path = "src"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

[tool.django-stubs]
django_settings_module = "_core.settings"
